Climate ensembles (#46)
* Set of ensembles

Allowing to use a set of climate ensembles as input to the model.

Correcting  boundary check issue where an input 0 for climate ensemble would not raise any error, misleadingly using climate ensemble 1000. Now the climate ensemble inputs must be in [1; 1001].

* fix

fixing a typo in a for loop over the climate ensembles.